Initial commit - Spring Boot project setup 6 (JWT Authentication)

+++ src/main/java/com/example/employee_api/repository/UserRepository.java
@@ -2,8 +2,21 @@ package com.example.employee_api.repository;
 
 import com.example.employee_api.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.*;
+import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+
     Optional<User> findByUsername(String username);
+
 }
+
+
+//package com.example.employee_api.repository;
+//
+//import com.example.employee_api.entity.User;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//import java.util.*;
+//
+//public interface UserRepository extends JpaRepository<User, Long> {
+//    Optional<User> findByUsername(String username);
+//}
